OS X 10.4 SDK @ Xcode 2.5
http://adcdownload.apple.com/Developer_Tools/xcode_2.5_developer_tools/xcode25_8m2558_developerdvd.dmg

+++ MacOSX10.4u.sdk/System/Library/Frameworks/Kernel.framework/Versions/A/Headers/IOKit/graphics/IODisplay.h
@@ -208,7 +208,7 @@ protected:
     UInt32	fCurrentPowerState;
     SInt32	fMinBrightness;
     SInt32	fMaxBrightness;
-    UInt8	fMaxBrightnessLevel[kIODisplayNumPowerStates];
+    UInt16	fMaxBrightnessLevel[kIODisplayNumPowerStates];
 
 public:
     virtual IOService * probe( IOService *, SInt32 * );
